65 configuration page UI logic (#70)
* Added configuration screen stuff

* Removed Add A Combination from UI

Removed all references to the Add a Combination screen in the UI since we do not need it there anymore

* Added dropdown to Settings script

* Configuration screen done

Configuration screen UI is fully done it just needs to be added to the database

* Added configuration object

Added the creation of the configuration object when attempting to save

* Added reference to SQLite handler in settings controller and added new database function provided by omar

* Configurations can be added

Configurations can now fully be added to the database

* Configuration deletion working

The ability to delete configurations is now properly working

* changing active configuration

You can now change the active configuration being used

* Almost completed Configuration page

Added accessibility features, such as adding placeholder texts and showing what percentage the gamma slider is at. Also added more error/success messages.
Code can be cleaned up a it and I am pretty sure there are a few spelling mistakes, but other than that, it seems good

* Removed local database and added a return

Removed the local database on my end from the PR. Added a return statement inside one of the if statements

* Cleaned up controller code

Cleaned up the code in the controller of the settings page

* fix: ensure handler instance is released on test setup

* fix: make release instance static to access before creating db handler

* fix: add webCamTexture null check before disposing it

* fix: add conditional compilation to EditorTextHandler helper

* fix: update create gesture according to merge conflicts

* Removing random sln file

Removing an unneeded sln file

---------

Co-authored-by: Omar Nunez <menesiena@proton.me>