67 media pipe gesture recognition functionality (#81)
* Step 1: Draw Landmarks

Got the camera in the scene drawing the landmarks using the previous code. This is just meant to be a visual aid for gestures in the next step

* Landmarks Gestures Recognized

The landmarks from the HandLandmarkerRunner.cs were uncovered and the GestureRecognizer.cs was created as a placeholder to recognize simple gestures. It's a little janky, but its progress.

* Player Example Setup

Got a basic player capsule implemented to showcase the map inputs. The gestures invoke an action, but still got to figure out how to map the input to the keyboard. Spent way too long trying to display the gestures in the UI (treading issues)

* Using GestureInputController

Instead of the GestureEvent, the Gesture is being passed to the GestureInputController. There is a threading issue

* UnityMainThreadDispatcher

Using this script to enqueue the background thread's execution to allow the invoke of events from the mediapipe

Co-Authored-By: Ryan S <90662533+RyanSamii@users.noreply.github.com>

---------

Co-authored-by: Ryan S <90662533+RyanSamii@users.noreply.github.com>