Directional gesture recognition (#90)
* New Player Input Setup

Created different colour toggle functions for the different point up, down, left, and right direction for better distinction of directional gestures

* Direction and Angle Calculations

Methods were added to calculate the vector direction of the gesture as 'up', 'down', 'left', 'right' to the database. The calculations and framework for the angles were included as well

* Attempted to update GestureSqliteHandler

Might be missing something, but I cannot create a table for directions in the database.

* Update Sqlite Handler and Reading Saved Directions

Modified the main scene, and updated the sqlite handler to save the directions to the database. Gesture Recognizer checks if the detected gesture is in the same direction before checking for a match.

* fix: properly delete all gesture data when deleted from ui

* feat: normalized gesture input mapping

* fix: stop ui window from auto-opening

This prevents the MainEditorWindow from opening when a scene is ran, or when the project is opened/reloaded. The window does not need to be initialized on load.

* feat: add camera customization through HandLandmarkerRunner

Also fixes test cases to add a gesture properly

* fix direction normalization

Also adds Gesture Recognizer prefab

* Improved Prefab and Display Gesture Names

The sample scene is only using the Gesture Recognizer prefab and the prefabs includes text displaying the name of each saved gesture

* fix: adjust gesture input controller tests after normalization changes

---------

Co-authored-by: Omar <menesiena@proton.me>